@@ -3,7 +3,7 @@
   "type": "library",
   "description": "Berlioz Framework Core",
   "license": "MIT",
-  "homepage": "http://getberlioz.com",
+  "homepage": "https://getberlioz.com",
   "authors": [
     {
       "name": "Ronan Giron",
@@ -21,30 +21,32 @@
   "autoload-dev": {
     "psr-4": {
       "Berlioz\\Core\\Tests\\": "tests/",
-      "Berlioz\\Core\\TestProject\\": "tests/_envTest/src/"
+      "Berlioz\\Core\\TestProject\\": "tests_env/src/"
     },
     "exclude-from-classmap": [
-      "/tests/_envTest/"
+      "/tests_env/src/"
     ]
   },
+  "minimum-stability": "beta",
+  "prefer-stable": true,
   "require": {
-    "php": "^7.1 || ^8.0",
+    "php": "^8.0",
     "ext-intl": "*",
-    "ext-json": "*",
     "ext-mbstring": "*",
     "ext-zlib": "*",
-    "berlioz/config": "^1.0",
-    "berlioz/helpers": "^1.0",
-    "berlioz/service-container": "^1.0",
-    "psr/log": "^1.0.0",
+    "berlioz/config": "^2.0",
+    "berlioz/event-manager": "^1.0",
+    "berlioz/helpers": "^1.2",
+    "berlioz/service-container": "^2.0",
+    "league/flysystem": "^2.1",
     "psr/simple-cache": "^1.0.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7.5 || ^8.0 || ^9.0"
+    "phpunit/phpunit": "^9.3"
   },
   "provide": {
     "psr/container-implementation": "^1.0",
-    "psr/log-implementation": "^1.0",
+    "psr/event-dispatcher-implementation": "^1.0",
     "psr/simple-cache-implementation": "^1.0"
   }
 }
