@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -14,69 +14,56 @@ declare(strict_types=1);
 
 namespace Berlioz\Core\Asset;
 
-use Berlioz\Config\ConfigAwareInterface;
-use Berlioz\Config\ConfigAwareTrait;
-use Berlioz\Config\ConfigInterface;
-use Berlioz\Config\Exception\ConfigException;
 use Berlioz\Core\Exception\AssetException;
 
 /**
  * Class Assets.
- *
- * @package Berlioz\Core\Asset
  */
-class Assets implements ConfigAwareInterface
+class Assets
 {
-    use ConfigAwareTrait;
-
-    /** @var Manifest Manifest */
-    private $manifest;
-    /** @var EntryPoints Entry points file */
-    private $entryPoints;
+    private ?Manifest $manifest = null;
+    private ?EntryPoints $entryPoints = null;
 
     /**
      * Assets constructor.
      *
-     * @param ConfigInterface $config
+     * @param string|null $manifestFile
+     * @param string|null $entryPointsFile
+     * @param string|null $entryPointsKey
+     *
+     * @throws AssetException
      */
-    public function __construct(ConfigInterface $config)
-    {
-        $this->setConfig($config);
+    public function __construct(
+        ?string $manifestFile = null,
+        ?string $entryPointsFile = null,
+        ?string $entryPointsKey = null,
+    ) {
+        if (null !== $manifestFile) {
+            $this->manifest = new Manifest($manifestFile);
+        }
+
+        if (null !== $entryPointsFile) {
+            $this->entryPoints = new EntryPoints($entryPointsFile, $entryPointsKey);
+        }
     }
 
     /**
      * Get manifest.
      *
-     * @return Manifest
-     * @throws ConfigException
-     * @throws AssetException
+     * @return Manifest|null
      */
-    public function getManifest(): Manifest
+    public function getManifest(): ?Manifest
     {
-        if (null === $this->manifest) {
-            $this->manifest = new Manifest($this->getConfig()->get('berlioz.assets.manifest'));
-        }
-
         return $this->manifest;
     }
 
     /**
      * Get entry points.
      *
-     * @return EntryPoints
-     * @throws ConfigException
-     * @throws AssetException
+     * @return EntryPoints|null
      */
-    public function getEntryPoints(): EntryPoints
+    public function getEntryPoints(): ?EntryPoints
     {
-        if (null === $this->entryPoints) {
-            $this->entryPoints =
-                new EntryPoints(
-                    $this->getConfig()->get('berlioz.assets.entrypoints'),
-                    $this->getConfig()->get('berlioz.assets.entrypoints_key')
-                );
-        }
-
         return $this->entryPoints;
     }
 }
