@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -18,8 +18,6 @@ use Berlioz\Core\Exception\AssetException;
 
 /**
  * Class Manifest.
- *
- * @package Berlioz\Core\Asset
  */
 class Manifest extends JsonAsset
 {
