@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,6 +16,9 @@ namespace Berlioz\Core\Cache;
 
 use Berlioz\Core\Exception\InvalidArgumentCacheException;
 
+/**
+ * Trait KeyControlTrait.
+ */
 trait KeyControlTrait
 {
     /**
