@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,8 +16,6 @@ namespace Berlioz\Core;
 
 /**
  * Interface CoreAwareInterface.
- *
- * @package Berlioz\Core
  */
 interface CoreAwareInterface
 {
@@ -35,5 +33,5 @@ interface CoreAwareInterface
      *
      * @return static
      */
-    public function setCore(Core $core);
+    public function setCore(Core $core): static;
 }
