@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -16,11 +16,16 @@ namespace Berlioz\Core\Directories;
 
 /**
  * Interface DirectoriesInterface.
- *
- * @package Berlioz\Core\Directories
  */
 interface DirectoriesInterface
 {
+    /**
+     * Get array copy.
+     *
+     * @return array
+     */
+    public function getArrayCopy(): array;
+
     /**
      * Get working directory.
      *
