@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -20,33 +20,47 @@ class EntryPointsTest extends TestCase
 {
     public function testTarget()
     {
-        $entryPoints = new EntryPoints(__DIR__ . '/../_files/entrypoints.target.json', 'entrypoints');
+        $entryPoints = new EntryPoints(__DIR__ . '/files/entrypoints.target.json', 'entrypoints');
         $this->assertInstanceOf(EntryPoints::class, $entryPoints);
-        $this->assertEquals(['/assets/js/test.12345678.js',
-                             '/assets/js/test2.12345678.js',
-                             '/assets/js/vendor.12345678.js'],
-                            $entryPoints->get('test', 'js'));
+        $this->assertEquals(
+            [
+                '/assets/js/test.12345678.js',
+                '/assets/js/test2.12345678.js',
+                '/assets/js/vendor.12345678.js'
+            ],
+            $entryPoints->get('test', 'js')
+        );
     }
 
     public function testBadTarget()
     {
         $this->expectException(AssetException::class);
-        $entryPoints = new EntryPoints(__DIR__ . '/../_files/entrypoints.json', 'entrypoints');
+        new EntryPoints(__DIR__ . '/files/entrypoints.json', 'entrypoints');
     }
 
     public function testGet()
     {
-        $entryPoints = new EntryPoints(__DIR__ . '/../_files/entrypoints.json');
+        $entryPoints = new EntryPoints(__DIR__ . '/files/entrypoints.json');
 
-        $this->assertEquals(['/assets/js/test.12345678.js',
-                             '/assets/js/test2.12345678.js',
-                             '/assets/js/vendor.12345678.js'],
-                            $entryPoints->get('test', 'js'));
-        $this->assertEquals(['js'  => ['/assets/js/test.12345678.js',
-                                       '/assets/js/test2.12345678.js',
-                                       '/assets/js/vendor.12345678.js'],
-                             'css' => ['/assets/css/test.12345678.css']],
-                            $entryPoints->get('test'));
+        $this->assertEquals(
+            [
+                '/assets/js/test.12345678.js',
+                '/assets/js/test2.12345678.js',
+                '/assets/js/vendor.12345678.js'
+            ],
+            $entryPoints->get('test', 'js')
+        );
+        $this->assertEquals(
+            [
+                'js' => [
+                    '/assets/js/test.12345678.js',
+                    '/assets/js/test2.12345678.js',
+                    '/assets/js/vendor.12345678.js'
+                ],
+                'css' => ['/assets/css/test.12345678.css']
+            ],
+            $entryPoints->get('test')
+        );
         $this->assertEquals([], $entryPoints->get('foo', 'js'));
         $this->assertEquals([], $entryPoints->get('test', 'bar'));
     }
