@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -18,6 +18,6 @@ class FakeDefaultDirectories extends DefaultDirectories
 {
     protected function getLibraryDirectory(): string
     {
-        return realpath(__DIR__ . '/../_envTest/vendor/berlioz/core');
+        return realpath(__DIR__ . '/../../tests_env/vendor/berlioz/core');
     }
 }
