@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -12,12 +12,14 @@
 
 namespace Berlioz\Core\Tests\Package;
 
+use Berlioz\Config\Adapter\ArrayAdapter;
+use Berlioz\Config\ConfigInterface;
 use Berlioz\Core\Package\AbstractPackage;
 
 class FakePackage3 extends AbstractPackage
 {
-    public static function config()
+    public static function config(): ?ConfigInterface
     {
-        return ['package3' => ['foo' => 'bar']];
+        return new ArrayAdapter(['package3' => ['foo' => 'bar']]);
     }
 }
