@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -12,14 +12,12 @@
 
 namespace Berlioz\Core\Tests\Package;
 
-use Berlioz\Config\JsonConfig;
-use Berlioz\Core\Composer;
 use Berlioz\Core\Core;
-use Berlioz\Core\Exception\BerliozException;
 use Berlioz\Core\Exception\PackageException;
 use Berlioz\Core\Package\PackageSet;
 use Berlioz\Core\Tests\Directories\FakeDefaultDirectories;
 use PHPUnit\Framework\TestCase;
+use stdClass;
 
 class PackageSetTest extends TestCase
 {
@@ -31,7 +29,7 @@ class PackageSetTest extends TestCase
         $this->assertCount(1, $packageSet->getPackages());
 
         $this->expectException(PackageException::class);
-        $packageSet->addPackage(\stdClass::class);
+        $packageSet->addPackage(stdClass::class);
     }
 
     public function testConfig()
@@ -41,21 +39,11 @@ class PackageSetTest extends TestCase
         $packageSet->addPackage(FakePackage2::class);
         $packageSet->addPackage(FakePackage3::class);
 
-        $packageSet->config($config = new JsonConfig('{}'));
+        $config = $packageSet->config();
 
         $this->assertEquals('qux', $config->get('package1'));
         $this->assertEquals('qux', $config->get('package2'));
         $this->assertEquals('bar', $config->get('package3.foo'));
-    }
-
-    public function testConfigError()
-    {
-        $packageSet = new PackageSet();
-        $packageSet->addPackage(FakePackageBad::class);
-
-        $this->expectException(BerliozException::class);
-
-        $packageSet->config($config = new JsonConfig('{}'));
     }
 
     public function testRegister()
@@ -64,12 +52,12 @@ class PackageSetTest extends TestCase
 
         $packageSet = new PackageSet();
         $packageSet->addPackage(FakePackage1::class);
-        $packageSet->register($core->getServiceContainer()->getInstantiator());
+        $packageSet->register($core->getContainer());
 
-        $this->assertTrue($core->getServiceContainer()->has('date'));
+        $this->assertTrue($core->getContainer()->has('date'));
     }
 
-    public function testInit()
+    public function testBoot()
     {
         $core = new Core(new FakeDefaultDirectories());
 
@@ -78,40 +66,39 @@ class PackageSetTest extends TestCase
 
         FakePackage1::$foo = false;
 
-        $packageSet->init($core->getServiceContainer()->getInstantiator());
+        $packageSet->boot($core);
 
         $this->assertTrue(FakePackage1::$foo);
     }
 
     public function testSerialization()
     {
-        $composer = new Composer(__DIR__ . '/../_envTest/composer.json');
         $packageSet = new PackageSet();
-        $packageSet->addPackagesFromComposer($composer);
+        $packageSet->addPackage(FakePackage1::class);
 
         $packageSet2 = unserialize(serialize($packageSet));
         $this->assertEquals($packageSet->getPackages(), $packageSet2->getPackages());
     }
 
-    public function testAddPackagesFromConfig()
-    {
-        $config = new JsonConfig(
-            '{"packages": ["Berlioz\\\\Core\\\\Tests\\\\Package\\\\FakePackage1", "Berlioz\\\\Core\\\\Tests\\\\Package\\\\FakePackage2"]}'
-        );
-        $packageSet = new PackageSet();
-        $packageSet->addPackagesFromConfig($config);
-
-        $this->assertCount(2, $packageSet->getPackages());
-        $this->assertEquals(["Berlioz\\Core\\Tests\\Package\\FakePackage1","Berlioz\\Core\\Tests\\Package\\FakePackage2"], $packageSet->getPackages());
-    }
-
-    public function testAddPackagesFromComposer()
-    {
-        $composer = new Composer(__DIR__ . '/../_envTest/composer.json');
-        $packageSet = new PackageSet();
-        $packageSet->addPackagesFromComposer($composer);
-
-        $this->assertCount(2, $packageSet->getPackages());
-        $this->assertEquals(["Berlioz\\Core\\Tests\\Package\\FakePackage1","Berlioz\\Core\\Tests\\Package\\FakePackage2"], $packageSet->getPackages());
-    }
+//    public function testAddPackagesFromConfig()
+//    {
+//        $config = new JsonConfig(
+//            '{"packages": ["Berlioz\\\\Core\\\\Tests\\\\Package\\\\FakePackage1", "Berlioz\\\\Core\\\\Tests\\\\Package\\\\FakePackage2"]}'
+//        );
+//        $packageSet = new PackageSet();
+//        $packageSet->addPackagesFromConfig($config);
+//
+//        $this->assertCount(2, $packageSet->getPackages());
+//        $this->assertEquals(["Berlioz\\Core\\Tests\\Package\\FakePackage1","Berlioz\\Core\\Tests\\Package\\FakePackage2"], $packageSet->getPackages());
+//    }
+//
+//    public function testAddPackagesFromComposer()
+//    {
+//        $composer = new Composer(__DIR__ . '/../../tests_env/composer.json');
+//        $packageSet = new PackageSet();
+//        $packageSet->addPackagesFromComposer($composer);
+//
+//        $this->assertCount(2, $packageSet->getPackages());
+//        $this->assertEquals(["Berlioz\\Core\\Tests\\Package\\FakePackage1","Berlioz\\Core\\Tests\\Package\\FakePackage2"], $packageSet->getPackages());
+//    }
 }
