@@ -1,16 +1,16 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2020 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code, to the root.
  */
 
-putenv(sprintf('DOCUMENT_ROOT=%s', realpath(__DIR__ . '/_envTest/public')));
+putenv(sprintf('DOCUMENT_ROOT=%s', realpath(__DIR__ . '/../tests_env/public')));
 chdir(getenv('DOCUMENT_ROOT'));
 
 $file = __DIR__ . '/../vendor/autoload.php';
